Scout Command refactoring and Paths command
- Refactored ScoutCommand into a protocol to offer centralised logic
- Added new paths command to list paths in the data
