+++ Sources/Scout/Implementations/XML/PathExplorerXML+Paths.swift
@@ -1,0 +1,11 @@
+//
+// Scout
+// Copyright (c) Alexis Bridoux 2020
+// MIT license, see LICENSE file for details
+
+extension PathExplorerXML {
+
+    public func getPaths(for filter: PathFilter?) -> [Path] {
+        return []
+    }
+}
