+++ Sources/ScoutCLT/Main/main.swift
@@ -3,4 +3,4 @@
 // Copyright (c) Alexis Bridoux 2020
 // MIT license, see LICENSE file for details
 
-ScoutCommand.main()
+ScoutMainCommand.main()
